requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
psycopg2-binary==2.9.9
pydantic==2.11.3
pydantic-settings==2.4.0
